requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
betterproto @ git+https://github.com/danielgtaylor/python-betterproto@1f88b67eeb9871d33da154fd2c859b9d1aed62c1
black==24.4.2
    # via betterproto
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via black
grpclib==0.4.7
    # via betterproto
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via grpclib
hpack==4.0.0
    # via h2
httpcore==1.0.5
    # via httpx
httpx==0.27.0
hyperframe==6.0.1
    # via h2
idna==3.7
    # via
    #   anyio
    #   httpx
isort==5.13.2
    # via betterproto
jinja2==3.1.4
    # via betterproto
markupsafe==2.1.5
    # via jinja2
multidict==6.0.5
    # via grpclib
mypy-extensions==1.0.0
    # via black
packaging==24.0
    # via black
paho-mqtt==2.1.0
pathspec==0.12.1
    # via black
platformdirs==4.2.2
    # via black
protobuf==5.27.0
pydantic==2.7.3
pydantic-core==2.18.4
    # via pydantic
python-dateutil==2.9.0.post0
    # via betterproto
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
typing-extensions==4.12.1
    # via
    #   betterproto
    #   pydantic
    #   pydantic-core
